+++ src/Build.UnitTests/EvaluationProfiler_Tests.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
@@ -12,8 +11,6 @@ using Microsoft.Build.Execution;
 using Microsoft.Build.Framework;
 using Microsoft.Build.Framework.Profiler;
 using Microsoft.Build.Logging;
-using Microsoft.Build.Shared;
-using Microsoft.Build.Utilities;
 using Xunit;
 using Xunit.Abstractions;
 using static Microsoft.Build.UnitTests.ObjectModelHelpers;
@@ -144,7 +141,7 @@ namespace Microsoft.Build.Engine.UnitTests
         {
             var content = CleanupFileContents(projectContent);
 
-            var profilerLogger = new ProfilerLogger("dummyFile");
+            var profilerLogger = ProfilerLogger.CreateForTesting();
             var parameters = new BuildParameters
             {
                 ShutdownInProcNodeOnBuildFinish = true,
